Move video related classes into video package
