+++ webrtc-kmp/src/iosMain/kotlin/com/shepeliev/webrtckmp/video/VideoCapturerController.kt
@@ -1,4 +1,6 @@
-package com.shepeliev.webrtckmp
+package com.shepeliev.webrtckmp.video
+
+import com.shepeliev.webrtckmp.MediaTrackSettings
 
 internal interface VideoCapturerController {
     val settings: MediaTrackSettings
